requirements.txt
flask
python-box
pyyaml
cherrypy
Flask-Session
reusables
PyJWT
appdirs
schema
pytest